added module for reading admins, streets, pois from osm
