+++ src/osm_reader/osm_admin_reader.rs
@@ -1,0 +1,140 @@
+// Copyright © 2016, Canal TP and/or its affiliates. All rights reserved.
+//
+// This file is part of Navitia,
+//     the software to build cool stuff with public transport.
+//
+// Hope you'll enjoy and contribute to this project,
+//     powered by Canal TP (www.canaltp.fr).
+// Help us simplify mobility and open public transport:
+//     a non ending quest to the responsive locomotion way of traveling!
+//
+// LICENCE: This program is free software; you can redistribute it
+// and/or modify it under the terms of the GNU Affero General Public
+// License as published by the Free Software Foundation, either
+// version 3 of the License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful, but
+// WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public
+// License along with this program. If not, see
+// <http://www.gnu.org/licenses/>.
+//
+// Stay tuned using
+// twitter @navitia
+// IRC #navitia on freenode
+// https://groups.google.com/d/forum/navitia
+// www.navitia.io
+
+extern crate osmpbfreader;
+extern crate mimir;
+
+use ::boundaries::{build_boundary, make_centroid};
+use std::cell::Cell;
+use std::collections::BTreeSet;
+use std::rc::Rc;
+use super::{OsmPbfReader, AdminsVec};
+
+#[derive(Debug)]
+pub struct AdminMatcher {
+    admin_levels: BTreeSet<u32>,
+}
+
+impl AdminMatcher {
+    pub fn new(levels: BTreeSet<u32>) -> AdminMatcher {
+        AdminMatcher { admin_levels: levels }
+    }
+
+    pub fn is_admin(&self, obj: &osmpbfreader::OsmObj) -> bool {
+        match *obj {
+            osmpbfreader::OsmObj::Relation(ref rel) => {
+                rel.tags.get("boundary").map_or(false, |v| v == "administrative") &&
+                rel.tags.get("admin_level").map_or(false, |lvl| {
+                    self.admin_levels.contains(&lvl.parse::<u32>().unwrap_or(0))
+                })
+            }
+            _ => false,
+        }
+    }
+}
+
+pub fn administrative_regions(pbf: &mut OsmPbfReader, levels: BTreeSet<u32>) -> AdminsVec {
+    let mut administrative_regions = AdminsVec::new();
+    let matcher = AdminMatcher::new(levels);
+    info!("reading pbf...");
+    let objects = osmpbfreader::get_objs_and_deps(pbf, |o| matcher.is_admin(o)).unwrap();
+    info!("reading pbf done.");
+    // load administratives regions
+    for (_, obj) in &objects {
+        if !matcher.is_admin(&obj) {
+            continue;
+        }
+        if let &osmpbfreader::OsmObj::Relation(ref relation) = obj {
+            let level = relation.tags
+                .get("admin_level")
+                .and_then(|s| s.parse().ok());
+            let level = match level {
+                None => {
+                    info!("invalid admin_level for relation {}: admin_level {:?}",
+                          relation.id,
+                          relation.tags.get("admin_level"));
+                    continue;
+                }
+                Some(l) => l,
+            };
+            // administrative region with name ?
+            let name = match relation.tags.get("name") {
+                Some(val) => val,
+                None => {
+                    warn!("adminstrative region without name for relation {}:  admin_level {} \
+                           ignored.",
+                          relation.id,
+                          level);
+                    continue;
+                }
+            };
+
+            // admininstrative region without coordinates
+            let coord_center = relation.refs
+                .iter()
+                .find(|rf| rf.role == "admin_centre")
+                .and_then(|r| {
+                    objects.get(&r.member).and_then(|value| {
+                        match value {
+                            &osmpbfreader::OsmObj::Node(ref node) => {
+                                Some(mimir::Coord::new(node.lat, node.lon))
+                            }
+                            _ => None,
+                        }
+                    })
+                });
+            let (admin_id, insee_id) = match relation.tags.get("ref:INSEE") {
+                Some(val) => {
+                    (format!("admin:fr:{}", val.trim_left_matches('0')), val.trim_left_matches('0'))
+                }
+                None => (format!("admin:osm:{}", relation.id), ""),
+            };
+
+            let zip_code = relation.tags
+                .get("addr:postcode")
+                .or_else(|| relation.tags.get("postal_code"))
+                .map(|val| &val[..])
+                .unwrap_or("");
+            let boundary = build_boundary(&relation, &objects);
+            let admin = mimir::Admin {
+                id: admin_id,
+                insee: insee_id.to_string(),
+                level: level,
+                label: name.to_string(),
+                zip_codes: zip_code.split(';').map(|s| s.to_string()).collect(),
+                weight: Cell::new(0),
+                coord: coord_center.unwrap_or_else(|| make_centroid(&boundary)),
+                boundary: boundary,
+            };
+            administrative_regions.push(Rc::new(admin));
+        }
+    }
+    return administrative_regions;
+}
